task 0 fixed

+++ user_authentication_service/user.py
@@ -26,5 +26,5 @@ class User(Base):
     id = Column(Integer, primary_key=True)
     email = Column(String(250), nullable=False)
     hashed_password = Column(String(250), nullable=False)
-    session_id = Column(Integer(250), primary_key=True)
+    session_id = Column(Integer, primary_key=True)
     reset_token = Column(String(250), nullable=True)
